requirements.txt
python-telegram-bot==20.3
torch>=2.0.0
torchvision>=0.15.1
Pillow>=9.5.0
requests>=2.31.0
scikit-learn>=1.2.2
pandas>=2.0.2
numpy